file-processing-service/requirements.txt
nats-py
minio
pandas
libmagic
python-magic
openpyxl
python-dotenv
watchdog